Fix Error in Python 3 Regarding Type Conversion
When you run this, line 170 typically gives an error in python3 saying that it cannot convert bytes to str.  This is because of type conversion differences in the languages.  I have proposed a solution.  When running `print(z)`, it gives the same value, proving that this change works.

Hopefully you agree with this change!  It fixed the problem for me.